Rename expand -> expansions
