+++ packages/ui/package.json
@@ -17,7 +17,7 @@
     ".": "./dist/public-api.js",
     "./dist/*": "./dist/*",
     "./cli": "./dist/cli/public-api.js",
-    "./cli-expansion": "./dist/cli/expand.js"
+    "./cli-expansion": "./dist/cli/expansion.js"
   },
   "files": [
     "dist",
